#161617593 Validation bug fixes (#5)
* [Bug #161617593] Fix all validation bugs

* [Bug #161617593] Update readme

* [Bug #161617593] Pep8 autoformat

* [Bug #161617593] Pep8 autoformat

* [Bug #161617593] Code refactor

* [Bug #161617593] Update readme

+++ app/api/v2/views/products.py
@@ -34,29 +34,30 @@ class Products(Resource):
         """
         Add a product to the manager
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         json_data = request.get_json(force=True)
         res = product_validator(json_data)
         if not res:
             store_id = get_store_id(get_jwt_identity())
             cur.execute(
-                "SELECT * FROM products WHERE name='{}';".format(json_data['name']))
+                "SELECT * FROM products WHERE name='{}';".format(json_data['name'].lower()))
             product = cur.fetchone()
             if product and product[1] == store_id:
                 msg = 'Product already exists.Update product inventory instead'
-                return {"status":"Failed!","message":msg},409
-            cat_name ='Category-not-set'
-            new_pro = Product(store_id, json_data['name'],
-                            json_data['inventory'],
-                            json_data['price'],
-                            cat_name)
+                return {"status": "Failed!", "message": msg}, 409
+            cat_name = 'Category-not-set'
+            new_pro = Product(store_id, json_data['name'].lower(),
+                              json_data['inventory'],
+                              json_data['price'],
+                              cat_name)
             new_pro.add_product()
             res = new_pro.json_dump()
             res = {"status": "Success!",
-                "message": "Product successfully added", "data": res}, 201
+                   "message": "Product successfully added", "data": res}, 201
         return res
-
-
-
 
     @v2.doc(security='apikey')
     @jwt_required
@@ -64,6 +65,10 @@ class Products(Resource):
         """
         Get all products
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         store_id = get_store_id(get_jwt_identity())
         cur.execute(
             "SELECT * FROM products WHERE store_id={};".format(store_id))
@@ -78,7 +83,8 @@ class Products(Resource):
                         'added_at': p[6]}
             all_products.append(format_p)
         if len(all_products) < 1:
-            res= {"status":"Failed!","message":"There are no products at the moment"},404
+            res = {"status": "Failed!",
+                   "message": "There are no products at the moment"}, 404
             return res
         return {"status": "Success!", "products": all_products}, 200
 
@@ -91,12 +97,16 @@ class ProductDetail(Resource):
         """
         Get single product
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         cur.execute("SELECT * FROM products WHERE id={};".format(id))
         product = cur.fetchone()
         store_id = get_store_id(get_jwt_identity())
         if not product or product[1] != store_id:
             msg = 'Product does not exist'
-            return {"status":"Failed!","message":msg},404
+            return {"status": "Failed!", "message": msg}, 404
         format_p = {
             "product_name": product[2],
             "inventory": product[3],
@@ -113,6 +123,10 @@ class ProductDetail(Resource):
         """
         Update a product
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         json_data = request.get_json(force=True)
         res = product_update_validator(json_data)
         if not res:
@@ -120,16 +134,22 @@ class ProductDetail(Resource):
             product = cur.fetchone()
             store_id = get_store_id(get_jwt_identity())
             if not product or product[1] != store_id:
-                return {"status":"Failed!","message": 'Product does not exist'}, 404
+                return {"status": "Failed!",
+                        "message": 'Product does not exist'}, 404
             name = product[2]
             inventory = product[3]
             price = product[4]
             if 'name' in json_data:
-                name = json_data['name']
+                name = json_data['name'].lower()
             if 'inventory' in json_data:
                 inventory = json_data['inventory']
             if 'price' in json_data:
                 price = json_data['price']
+            cur.execute("SELECT * FROM products WHERE name='{}';".format(name))
+            product_check = cur.fetchone()
+            if product_check:
+                msg = 'That product already exists'
+                return {"status": "Failed!", "message": msg}, 406
             cur.execute("UPDATE products SET name='{}',inventory='{}',price='{}'\
             WHERE id ={}".format(name, inventory, price, id))
             conn.commit()
@@ -152,12 +172,16 @@ class ProductDetail(Resource):
         """
         Delete a product
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         cur.execute("SELECT * FROM products WHERE id={};".format(id))
         product = cur.fetchone()
         store_id = get_store_id(get_jwt_identity())
         if not product or product[1] != store_id:
             msg = 'Product does not exist'
-            return {"status":"Failed!","message":msg},404
+            return {"status": "Failed!", "message": msg}, 404
         cur.execute("DELETE FROM products WHERE id={};".format(id))
         conn.commit()
         format_p = {
@@ -174,6 +198,10 @@ class ProductDetail(Resource):
         """
         Add a product to cart
         """
+        current_user = get_jwt_identity()
+        if current_user is None:
+            msg = 'Please login to access to access this resource'
+            return {"status": "Failed!", "message": msg}, 400
         json_data = request.get_json(force=True)
         res = sales_validator(json_data)
         if not res:
@@ -183,12 +211,12 @@ class ProductDetail(Resource):
             store_id = get_store_id(get_jwt_identity())
             if not product or product[1] != store_id:
                 msg = 'Product does not exist'
-                return {"status":"Failed!","message":msg},404
+                return {"status": "Failed!", "message": msg}, 404
             product_name = product[2]
             if product[3] < int(number):
                 msg = 'There are only {0} {1} available'.format(
                     product[3], product_name)
-                return {"status":"Failed!","message":msg},400
+                return {"status": "Failed!", "message": msg}, 400
             amount = number * product[4]
             seller = get_user_by_email(get_jwt_identity())
             seller_id = seller[0]
@@ -199,5 +227,8 @@ class ProductDetail(Resource):
             cur.execute(
                 "UPDATE products SET inventory={0} WHERE id ={1}".format(
                     new_inventory, id))
-            res = {"status":"Success!","message":"Added to cart", "product": res}
+            res = {
+                "status": "Success!",
+                "message": "Added to cart",
+                "product": res}
         return res
